dashboard/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
hopsworks>=3.0.0
hsml>=3.0.0
joblib>=1.3.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
